requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
pydantic==2.11.7
PyPDF2>=3.0.0
pdfminer.six>=20250506
sentence-transformers==4.1.0
langchain==0.3.26
langchain-community==0.3.26
pypdf==5.6.0
tiktoken==0.9.0
faiss-cpu==1.11.0
pymongo==4.13.0
openai==1.84.0
certifi==2025.6.15
fastapi-mcp==0.3.4
crewai==0.141.0
crewai-tools==0.47.1
argon2-cffi==25.1.0
streamlit==1.46.1
dateparser==1.2.2
docling==2.41.0
litellm==1.72.6
colorama==0.4.6
redis==6.2.0
streamlit-cookies-manager==0.2.0
mlflow==3.1.1
ragas==0.3.1
crewai-tools[mcp]
